Renamed Photos Component from Parallax Scroll
Renames the 'parallax-scroll' UI component to 'photos' for clearer text.
